move SubjectContext to Elements SDK directory
