+++ frontend/src/Elements/Element.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { inject } from './contexts/SubjectContext';
 
 import Consent from './components/Consent/Consent';
 import ForgottenRequest from './components/Erasure/ForgottenRequest';
@@ -9,15 +10,15 @@ import Objection from './components/Objection/Objection';
 import Rectification from './components/Rectification/Rectification';
 import Restriction from './components/Restriction/Restriction';
 
-const ELEMENT_TYPES = {
-  CONSENT: 'consent',
-  ERASURE: 'forgotten',
-  OBJECTION: 'objection',
-  RECTIFICATION: 'rectification',
-  DATA_EXPORT: 'export',
-  DATA_STATUS: 'data',
-  DATA_SHARE: 'share-data',
-  RESTRICTION: 'restriction'
+const ELEMENTS = {
+  consent: inject(Consent),
+  forgotten: inject(ForgottenRequest),
+  objection: inject(Objection),
+  rectification: inject(Rectification),
+  export: inject(ExportRequest),
+  data: inject(UserData),
+  'share-data': inject(ShareData),
+  restriction: inject(Restriction)
 };
 
 /**
@@ -25,25 +26,10 @@ const ELEMENT_TYPES = {
  */
 export default class Element {
   static create({ source, cg, ...options }) {
-    switch (source) {
-      case ELEMENT_TYPES.CONSENT:
-        return <Consent {...{ options, cg }} />;
-      case ELEMENT_TYPES.ERASURE:
-        return <ForgottenRequest {...{ options, cg }} />;
-      case ELEMENT_TYPES.DATA_EXPORT:
-        return <ExportRequest {...{ options, cg }} />;
-      case ELEMENT_TYPES.DATA_STATUS:
-        return <UserData {...{ options, cg }} />;
-      case ELEMENT_TYPES.DATA_SHARE:
-        return <ShareData options={options} />;
-      case ELEMENT_TYPES.OBJECTION:
-        return <Objection {...{ options, cg }} />;
-      case ELEMENT_TYPES.RECTIFICATION:
-        return <Rectification options={options} />;
-      case ELEMENT_TYPES.RESTRICTION:
-        return <Restriction {...{ options, cg }} />;
-      default:
-        throw new Error('Element was type not specified.');
+    const element = ELEMENTS[source];
+    if (!element) {
+      throw new Error('Element was type not specified.');
     }
+    return React.createElement(element, { options, cg });
   }
 }
